+++ src/services/database-split.js
@@ -1,10 +1,14 @@
 // temporary
 
+import dePopulate from './de-populate';
 import legacyPopulate from './legacy-populate';
+import populate from './populate';
 import serialize from './serialize';
 import softDelete from './soft-delete';
 
 export default {
+  dePopulate,
+  populate,
   serialize,
   legacyPopulate,
   softDelete
