+++ src/services/index.js
@@ -2,7 +2,6 @@
 module.exports = Object.assign({},
   require('./alter-data-split'),
   require('./query-params-split'),
-  require('./populate'),
   require('./database-split'),
   require('./validate'),
   require('./utilities'),
